+++ changelog.md
@@ -1,4 +1,6 @@
 
 # Changelog
 
+* Added Calculator component and form
+* Added loading of sass files from webpack
 * added VmWare-clarity https://vmware.github.io/clarity/index.html